Separated training and cross-validation scripts Copied train.py to new cross_validate.py and reverted to original training script. Common functions moved to src/utils and unused imports were removed. HDF5 handling class moved to src/data_pipeline. Updated configs.
