+++ src/data_pipeline.py
@@ -6,6 +6,10 @@ from pyFAI.detectors import Detector
 from pyFAI.geometry import Geometry
 from pyFAI.integrator.azimuthal import AzimuthalIntegrator
 from typing import cast
+import torch
+from torch.utils.data import Dataset
+import h5py
+from torchvision.transforms import Resize
 
 class CalibrantSim:
     """
@@ -91,3 +95,26 @@ class CalibrantSim:
 
         noisy_image = np.random.poisson(image * snr) / snr
         return noisy_image.astype(np.float32)
+    
+class HDF5Dataset(Dataset):
+    """PyTorch Dataset for loading data from an HDF5 file."""
+    def __init__(self, hdf5_path: str, group: str, image_size: int = 224):
+        self.file = h5py.File(hdf5_path, 'r')
+        self.group = self.file[group]
+        self.images = self.group['images'] #type: ignore
+        self.labels = self.group['labels'] #type: ignore
+        self.resize_transform = Resize((image_size, image_size), antialias=True)
+
+    def __len__(self):
+        return len(self.images) #type: ignore
+
+    def __getitem__(self, idx):
+        image = self.images[idx] #type: ignore
+        image_tensor = torch.from_numpy(image).unsqueeze(0).float()
+        image_tensor = self.resize_transform(image_tensor)
+        image_tensor = image_tensor.repeat(3, 1, 1)
+        label_tensor = torch.from_numpy(self.labels[idx]).float() #type: ignore
+        return image_tensor, label_tensor
+
+    def close(self):
+        self.file.close()
